Refactor Zed extension to use getPackageManager() API
Extract package management functionality from PositronZedRuntimeSession
into a separate ZedPackageManager class that implements the
LanguageRuntimePackageManager interface. This brings the Zed test
extension in line with the R and Python extensions which use the same
pattern.

- Create ZedPackageManager class with all package operations
- Add getPackageManager() method to session class
- Remove package methods from session class
